Fix wording in English and spacing in Spanish (#43)
As agreed, I changed the unclear cases of "pass" to "succeed" in
English, fixes #38

In looking through the translations, I happened to notice a spacing
issue in Spanish so I fixed that one as well - pro bono. 😉
It's good that there are those review issues for existing languages
because the sentence "Multiplier increases by 1 each round if you
pass/succeed" may need improvement in other languages as well. I know
German is 👌🏻 (clearly refers to winning instead of skipping) but I'm
unsure about the other languages.